+++ components/core/markdown/src/main/java/com/flipperdevices/core/markdown/ComposableMarkdown.kt
@@ -8,63 +8,68 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
-import com.flipperdevices.core.ui.theme.models.FlipperPallet
-import com.flipperdevices.core.ui.theme.models.FlipperTypography
 import com.mikepenz.markdown.compose.Markdown
 import com.mikepenz.markdown.model.MarkdownColors
 import com.mikepenz.markdown.model.MarkdownPadding
 import com.mikepenz.markdown.model.MarkdownTypography
 
 @Composable
-fun ComposableMarkdown(content: String, modifier: Modifier) {
+fun ComposableMarkdown(
+    modifier: Modifier = Modifier,
+    content: String,
+    typography: MarkdownTypography = markdownTypography(),
+    colors: MarkdownColors = markdownColors()
+) {
     Markdown(
         content = content,
-        colors = LocalPallet.current.toMarkdownColor(),
-        typography = LocalTypography.current.toMarkdownTypography(),
+        colors = colors,
+        typography = typography,
         padding = markdownPadding(),
         modifier = modifier
     )
 }
 
 @Composable
-private fun FlipperPallet.toMarkdownColor(): MarkdownColors {
+fun markdownColors(
+    backgroundCode: Color = LocalPallet.current.text20,
+    text: Color = LocalPallet.current.text100
+): MarkdownColors {
     return object : MarkdownColors {
-        override val backgroundCode: Color
-            get() = text20
-        override val text: Color
-            get() = text100
+        override val backgroundCode = backgroundCode
+        override val text = text
     }
 }
 
 @Composable
-private fun FlipperTypography.toMarkdownTypography(): MarkdownTypography {
+fun markdownTypography(
+    bulletStyle: TextStyle = LocalTypography.current.bodyR14,
+    codeStyle: TextStyle = LocalTypography.current.bodyR14,
+    h1Style: TextStyle = LocalTypography.current.bodySB14,
+    h2Style: TextStyle = LocalTypography.current.bodySB14,
+    h3Style: TextStyle = LocalTypography.current.bodySB14,
+    h4Style: TextStyle = LocalTypography.current.bodySB14,
+    h5Style: TextStyle = LocalTypography.current.bodySB14,
+    h6Style: TextStyle = LocalTypography.current.bodySB14,
+    listStyle: TextStyle = LocalTypography.current.bodyR14,
+    orderedStyle: TextStyle = LocalTypography.current.bodyR14,
+    paragraphStyle: TextStyle = LocalTypography.current.bodyR14,
+    quoteStyle: TextStyle = LocalTypography.current.bodyR14,
+    textStyle: TextStyle = LocalTypography.current.bodyR14
+): MarkdownTypography {
     return object : MarkdownTypography {
-        override val bullet: TextStyle
-            get() = bodyR14
-        override val code: TextStyle
-            get() = bodyR14
-        override val h1: TextStyle
-            get() = bodySB14
-        override val h2: TextStyle
-            get() = bodySB14
-        override val h3: TextStyle
-            get() = bodySB14
-        override val h4: TextStyle
-            get() = bodySB14
-        override val h5: TextStyle
-            get() = bodySB14
-        override val h6: TextStyle
-            get() = bodySB14
-        override val list: TextStyle
-            get() = bodyR14
-        override val ordered: TextStyle
-            get() = bodyR14
-        override val paragraph: TextStyle
-            get() = bodyR14
-        override val quote: TextStyle
-            get() = bodyR14
-        override val text: TextStyle
-            get() = bodyR14
+        override val bullet = bulletStyle
+        override val code = codeStyle
+        override val h1 = h1Style
+        override val h2 = h2Style
+        override val h3 = h3Style
+        override val h4 = h4Style
+        override val h5 = h5Style
+        override val h6 = h6Style
+        override val list = listStyle
+        override val ordered = orderedStyle
+        override val paragraph = paragraphStyle
+        override val quote = quoteStyle
+        override val text = textStyle
     }
 }