+++ components/updater/api/src/main/java/com/flipperdevices/updater/api/DownloaderApi.kt
@@ -3,6 +3,8 @@ package com.flipperdevices.updater.api
 import com.flipperdevices.updater.model.DistributionFile
 import com.flipperdevices.updater.model.DownloadProgress
 import com.flipperdevices.updater.model.FirmwareChannel
+import com.flipperdevices.updater.model.SubGhzProvisioningException
+import com.flipperdevices.updater.model.SubGhzProvisioningModel
 import com.flipperdevices.updater.model.VersionFiles
 import java.io.File
 import java.util.EnumMap
@@ -10,6 +12,9 @@ import kotlinx.coroutines.flow.Flow
 
 interface DownloaderApi {
     suspend fun getLatestVersion(): EnumMap<FirmwareChannel, VersionFiles>
+
+    @Throws(SubGhzProvisioningException::class)
+    suspend fun getSubGhzProvisioning(): SubGhzProvisioningModel
 
     fun download(
         distributionFile: DistributionFile,