+++ components/updater/api/src/main/java/com/flipperdevices/updater/model/UpdatingState.kt
@@ -13,9 +13,13 @@ sealed class UpdatingState(
 ) {
     object NotStarted : UpdatingState(true)
 
+    object FailedSubGhzProvisioning : UpdatingState(true)
+    object FailedOutdatedApp : UpdatingState(true)
     object FailedDownload : UpdatingState(true)
     object FailedPrepare : UpdatingState(true)
     object FailedUpload : UpdatingState(true)
+
+    object SubGhzProvisioning : UpdatingState(false)
 
     data class DownloadingFromNetwork(
         val percent: Float