+++ components/updater/screen/src/main/java/com/flipperdevices/updater/screen/composable/ComposableUpdateContent.kt
@@ -21,6 +21,7 @@ import com.flipperdevices.updater.screen.model.FailedReason
 import com.flipperdevices.updater.screen.model.UpdaterScreenState
 
 @Composable
+@Suppress("LongMethod", "ComplexMethod")
 fun ComposableUpdateContent(
     updaterScreenState: UpdaterScreenState,
     onRetry: () -> Unit
@@ -51,6 +52,12 @@ fun ComposableUpdateContent(
             iconId = DesignSystem.drawable.ic_globe,
             percent = updaterScreenState.percent
         )
+        is UpdaterScreenState.SubGhzProvisioning -> ComposableInProgressIndicator(
+            accentColor = LocalPallet.current.updateProgressGreen,
+            secondColor = LocalPallet.current.updateProgressBackgroundGreen,
+            iconId = DesignSystem.drawable.ic_globe,
+            percent = null
+        )
         is UpdaterScreenState.UploadOnFlipper -> ComposableInProgressIndicator(
             accentColor = LocalPallet.current.accentSecond,
             secondColor = LocalPallet.current.updateProgressBackgroundBlue,
@@ -72,6 +79,8 @@ fun ComposableUpdateContent(
         is UpdaterScreenState.Failed -> when (updaterScreenState.failedReason) {
             FailedReason.UPLOAD_ON_FLIPPER -> ComposableFailedUploadContent()
             FailedReason.DOWNLOAD_FROM_NETWORK -> ComposableFailedDownloadContent(onRetry)
+            FailedReason.OUTDATED_APP -> ComposableOutdatedApp()
+            FailedReason.FAILED_SUB_GHZ_PROVISIONING -> ComposableInternalFlashFailed()
         }
         UpdaterScreenState.Finish -> return
     }
@@ -100,6 +109,7 @@ private fun DescriptionUpdateText(
         is UpdaterScreenState.CancelingSynchronization -> R.string.update_stage_sync_canceling_desc
         UpdaterScreenState.CancelingUpdate -> R.string.update_stage_update_canceling_desc
         is UpdaterScreenState.DownloadingFromNetwork -> R.string.update_stage_downloading_desc
+        is UpdaterScreenState.SubGhzProvisioning -> R.string.update_stage_subghz_desc
         UpdaterScreenState.Finish -> R.string.update_stage_update_canceling_desc
         UpdaterScreenState.NotStarted -> R.string.update_stage_starting_desc
         UpdaterScreenState.Rebooting -> R.string.update_stage_rebooting_desc